backend/requirements.txt
# Flask Core
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Bcrypt==1.0.1

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
supabase==2.3.0

# Web Scraping
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.15.2
scrapy==2.11.0
lxml==4.9.3
facebook-scraper==0.2.59

# Machine Learning & NLP
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
transformers==4.35.2
torch==2.1.1
spacy==3.7.2

# Data Processing
python-dateutil==2.8.2
python-dotenv==1.0.0
tqdm==4.66.1

# Scheduling
APScheduler==3.11.1

# API & Validation
marshmallow==3.20.1
python-multipart==0.0.6

# Task Queue (optionnel pour scraping asynchrone)
celery==5.3.4
redis==5.0.1

# Monitoring & Logs
python-json-logger==2.0.7

# Utils
click==8.1.7
Werkzeug==3.0.1
